Fixed framebuffer
I was going to implement mailbox interface using structs but it turned out
that my approach generated to much code in comparison to simple array buffer approach that is presented
in tutorials. At this point the soon-to-be OS is very, very simple so I'm going to stick with basic array.
Also the less complex code is the less bugs it generates.